refactor(menu): Optimize menu tree query logic and controller code

+++ apps/api/src/modules/admin/system/menu/service.ts
@@ -41,17 +41,23 @@ export class Service {
   }
 
   async treeselect(uid, params) {
-    const query = this.rootRoleId === uid 
-    ? `select menu_id, menu_name, parent_id, order_num, path, component, query, is_frame, is_cache, menu_type, visible, status, ifnull(perms,'') as perms, icon, create_time 
-       from sys_menu order by parent_id, order_num`
-    : `select distinct m.menu_id, m.parent_id, m.menu_name, m.path, m.component, m.query, m.visible, m.status, ifnull(m.perms,'') as perms, m.is_frame, m.is_cache, m.menu_type, m.icon, m.order_num, m.create_time
-       from sys_menu m
-       left join sys_role_menu rm on m.menu_id = rm.menu_id
-       left join sys_user_role ur on rm.role_id = ur.role_id
-       left join sys_role ro on ur.role_id = ro.role_id
-       where ur.user_id = ${params.userId}
-       order by m.parent_id, m.order_num`;
-    const menuList: any = await prisma.$queryRaw`${query}`;
+    let menuList: any;
+    if (this.rootRoleId === uid) { // 如果是 root 用户，查询所有菜单
+      menuList = await prisma.$queryRaw`
+        select menu_id, menu_name, parent_id, order_num, path, component, query, is_frame, is_cache, menu_type, visible, status, ifnull(perms,'') as perms, icon, create_time
+        from sys_menu
+        order by parent_id, order_num`;
+    } else {
+      const targetUserId = params.userId || uid; // 如果不是 root 用户，根据用户ID查询其有权限的菜单
+      menuList = await prisma.$queryRaw`
+        select distinct m.menu_id, m.parent_id, m.menu_name, m.path, m.component, m.query, m.visible, m.status, ifnull(m.perms,'') as perms, m.is_frame, m.is_cache, m.menu_type, m.icon, m.order_num, m.create_time
+        from sys_menu m
+        left join sys_role_menu rm on m.menu_id = rm.menu_id
+        left join sys_user_role ur on rm.role_id = ur.role_id
+        left join sys_role ro on ur.role_id = ro.role_id
+        where ur.user_id = ${targetUserId}
+        order by m.parent_id, m.order_num`;
+    }
     const menuArr = menuList.map((item: any) => ({
       ...item,
       parentId: Number(item.parent_id),
